Implement Stannum::Entity.
