+++ spec/integration/entities/gadget_spec.rb
@@ -1,21 +1,20 @@
 # frozen_string_literal: true
 
-require 'support/structs/gizmo'
+require 'support/entities/gadget'
 
-# @note Integration spec for a subclassed Stannum::Struct.
-RSpec.describe Spec::Gizmo do
+# @note Integration spec for Stannum::Entity.
+RSpec.describe Spec::Gadget do
   shared_context 'when initialized with attribute values' do
     let(:attributes) do
       {
         name:        'Self-Sealing Stem Bolt',
         description: 'No one is sure what a self-sealing stem bolt is.',
-        quantity:    1_000,
-        size:        'Small'
+        quantity:    1_000
       }
     end
   end
 
-  subject(:gadget) { described_class.new(attributes) }
+  subject(:gadget) { described_class.new(**attributes) }
 
   let(:attributes) { {} }
 
@@ -43,7 +42,7 @@ RSpec.describe Spec::Gizmo do
     end
 
     describe '.each' do
-      let(:expected_keys) { %w[name description quantity size] }
+      let(:expected_keys) { %w[name description quantity] }
       let(:expected_values) do
         [
           {
@@ -60,11 +59,6 @@ RSpec.describe Spec::Gizmo do
             name:    'quantity',
             type:    'Integer',
             options: { default: 0, required: true }
-          },
-          {
-            name:    'size',
-            type:    'String',
-            options: { required: true }
           }
         ].map do |attributes|
           an_instance_of(Stannum::Attribute).and(
@@ -75,7 +69,7 @@ RSpec.describe Spec::Gizmo do
 
       it { expect(described_class.attributes.each).to be_a Enumerator }
 
-      it { expect(described_class.attributes.each.size).to be 4 }
+      it { expect(described_class.attributes.each.size).to be 3 }
 
       it 'should yield the attributes' do
         expect { |block| described_class.attributes.each(&block) }
@@ -99,21 +93,9 @@ RSpec.describe Spec::Gizmo do
             type:    'stannum.constraints.is_not_type'
           },
           {
-            data:    { required: true, type: String },
-            message: nil,
-            path:    [:size],
-            type:    'stannum.constraints.is_not_type'
-          },
-          {
             data:    {},
             message: nil,
             path:    [],
-            type:    'stannum.constraints.invalid'
-          },
-          {
-            data:    {},
-            message: nil,
-            path:    [:size],
             type:    'stannum.constraints.invalid'
           }
         ]
@@ -129,8 +111,7 @@ RSpec.describe Spec::Gizmo do
         {
           name:        'Self-Sealing Stem Bolt',
           description: :invalid_description,
-          quantity:    -10,
-          size:        'Lilliputian'
+          quantity:    -10
         }
       end
       let(:expected_errors) do
@@ -152,12 +133,6 @@ RSpec.describe Spec::Gizmo do
             message: nil,
             path:    [:quantity],
             type:    'stannum.constraints.invalid'
-          },
-          {
-            data:    {},
-            message: nil,
-            path:    [:size],
-            type:    'stannum.constraints.invalid'
           }
         ]
       end
@@ -172,8 +147,7 @@ RSpec.describe Spec::Gizmo do
         {
           name:        'Self-Sealing Stem Bolt',
           description: 'No one is sure what a self-sealing stem bolt is.',
-          quantity:    0,
-          size:        'Small'
+          quantity:    0
         }
       end
 
@@ -187,35 +161,19 @@ RSpec.describe Spec::Gizmo do
     it { expect(gadget).to respond_to(:[]).with(1).argument }
 
     describe 'with a string key' do
-      it { expect(gadget['size']).to be nil }
-    end
-
-    describe 'with a symbol key' do
-      it { expect(gadget[:size]).to be nil }
-    end
-
-    describe 'with a string key for a parent attribute' do
       it { expect(gadget['name']).to be nil }
     end
 
-    describe 'with a symbol key for a parent attribute' do
+    describe 'with a symbol key' do
       it { expect(gadget[:name]).to be nil }
     end
 
     wrap_context 'when initialized with attribute values' do
       describe 'with a string key' do
-        it { expect(gadget['size']).to be == attributes[:size] }
-      end
-
-      describe 'with a symbol key' do
-        it { expect(gadget[:size]).to be == attributes[:size] }
-      end
-
-      describe 'with a string key for a parent attributes' do
         it { expect(gadget['name']).to be == attributes[:name] }
       end
 
-      describe 'with a symbol key for a parent attribute' do
+      describe 'with a symbol key' do
         it { expect(gadget[:name]).to be == attributes[:name] }
       end
     end
@@ -242,29 +200,11 @@ RSpec.describe Spec::Gizmo do
   end
 
   describe '#[]=' do
-    let(:value) { 'Colossal' }
+    let(:value) { 'Inverse Chronoton Emitter' }
 
     it { expect(gadget).to respond_to(:[]=).with(2).arguments }
 
     describe 'with a string key' do
-      it 'should update the value' do
-        expect { gadget['size'] = value }
-          .to change(gadget, :size)
-          .to be == value
-      end
-    end
-
-    describe 'with a symbol key' do
-      it 'should update the value' do
-        expect { gadget[:size] = value }
-          .to change(gadget, :size)
-          .to be == value
-      end
-    end
-
-    describe 'with a string key for a parent attribute' do
-      let(:value) { 'Inverse Chronoton Emitter' }
-
       it 'should update the value' do
         expect { gadget['name'] = value }
           .to change(gadget, :name)
@@ -272,9 +212,7 @@ RSpec.describe Spec::Gizmo do
       end
     end
 
-    describe 'with a symbol key for a parent attribute' do
-      let(:value) { 'Inverse Chronoton Emitter' }
-
+    describe 'with a symbol key' do
       it 'should update the value' do
         expect { gadget[:name] = value }
           .to change(gadget, :name)
@@ -285,33 +223,13 @@ RSpec.describe Spec::Gizmo do
     wrap_context 'when initialized with attribute values' do
       describe 'with a string key' do
         it 'should update the value' do
-          expect { gadget['size'] = value }
-            .to change(gadget, :size)
-            .to be == value
-        end
-      end
-
-      describe 'with a symbol key' do
-        it 'should update the value' do
-          expect { gadget[:size] = value }
-            .to change(gadget, :size)
-            .to be == value
-        end
-      end
-
-      describe 'with a string key for a parent attribute' do
-        let(:value) { 'Inverse Chronoton Emitter' }
-
-        it 'should update the value' do
           expect { gadget['name'] = value }
             .to change(gadget, :name)
             .to be == value
         end
       end
 
-      describe 'with a symbol key for a parent attribute' do
-        let(:value) { 'Inverse Chronoton Emitter' }
-
+      describe 'with a symbol key' do
         it 'should update the value' do
           expect { gadget[:name] = value }
             .to change(gadget, :name)
@@ -382,8 +300,7 @@ RSpec.describe Spec::Gizmo do
       {
         'name'        => nil,
         'description' => description,
-        'quantity'    => 0,
-        'size'        => nil
+        'quantity'    => 0
       }
     end
 
@@ -416,8 +333,7 @@ RSpec.describe Spec::Gizmo do
       {
         'name'        => nil,
         'description' => nil,
-        'quantity'    => 0,
-        'size'        => nil
+        'quantity'    => 0
       }
     end
 
@@ -442,8 +358,7 @@ RSpec.describe Spec::Gizmo do
       {
         'name'        => nil,
         'description' => description,
-        'quantity'    => 0,
-        'size'        => nil
+        'quantity'    => 0
       }
     end
 
@@ -564,34 +479,6 @@ RSpec.describe Spec::Gizmo do
             .to change(gadget, :quantity)
             .to be 100
         end
-      end
-    end
-  end
-
-  describe '#size' do
-    include_examples 'should have reader', :size
-
-    wrap_context 'when initialized with attribute values' do
-      it { expect(gadget.size).to be == 'Small' }
-    end
-  end
-
-  describe '#size=' do
-    let(:value) { 'Colossal' }
-
-    include_examples 'should have writer', :size=
-
-    it 'should update the value' do
-      expect { gadget.size = value }
-        .to change(gadget, :size)
-        .to be == value
-    end
-
-    wrap_context 'when initialized with attribute values' do
-      it 'should update the value' do
-        expect { gadget.size = value }
-          .to change(gadget, :size)
-          .to be == value
       end
     end
   end
